Move stuff from `main` to `serialize`.
