+++ src/lib.rs
@@ -9,11 +9,11 @@ pub mod f_table;
 pub mod more_ops;
 pub mod node;
 pub mod number;
+pub mod object_cache;
 pub mod op_utils;
 pub mod reduction;
 pub mod run_program;
 pub mod runtime_dialect;
-pub mod serialization_cache;
 pub mod serialize;
 pub mod sha2;
 pub mod stack_cache;
