+++ src/object_cache.rs
